refactored out some colors, removed some unused icons, adjusted webpack config to ignore *.d.ts files as "unused", did a full audit of colors and added comments for number of files for each color to help us narrow things down, need to discuss with design. fixed some file names from style.less to styles.less to fit conventions. removed _inputs.less and put it directly into shared, removed 2 classes in shares that were unused.
