+++ packages/augur-ui/src/assets/styles/_buttons.less
@@ -22,7 +22,7 @@
   }
 
   &.imperative {
-    background-color: @color-error-text;
+    background-color: @color-error;
   }
 
   &.navigational {
@@ -94,7 +94,7 @@
   padding: 0.75rem 1.5rem;
 
   &:hover {
-    background: @color-primary-button-background-hover;
+    background: fade(@color-primary-action, 90%);
     border: 0;
     color: @color-secondary-action;
   }
