--- packages/augur-ui/src/assets/styles/_inputs.less
@@ -1,34 +1,0 @@
-label {
-  color: @color-secondary-text;
-  display: block;
-  margin-bottom: @size-10;
-
-  &[for] {
-    cursor: pointer;
-  }
-}
-
-input,
-textarea {
-  background-color: @color-secondary-text;
-  border: @size-1 solid transparent;
-  color: @color-theme-primary;
-  display: block;
-  font-family: @roboto;
-  padding: @size-12 @size-16;
-  width: 100%;
-
-  &:focus {
-    border: @size-1 solid @color-primary-light;
-  }
-
-  &:disabled {
-    background-color: @color-disabled-text-secondary;
-    cursor: default;
-  }
-
-  &::placeholder {
-    color: @color-alternate-contrast;
-    font-family: @roboto;
-  }
-}
