+++ packages/augur-ui/src/modules/account/components/status.styles.less
@@ -106,7 +106,7 @@
   > div:nth-of-type(4) {
     .mono-10;
 
-    color: @color-in-reporting;
+    color: @color-reporting;
     margin: 2px 0 0;
     text-align: center;
   }
@@ -142,7 +142,7 @@
   }
 
   > div:nth-of-type(4) {
-    color: @color-in-reporting;
+    color: @color-reporting;
     margin: @size-4 @size-36 0;
   }
 }
