+++ packages/augur-ui/src/modules/alerts/components/alert.styles.less
@@ -10,7 +10,7 @@
   text-overflow: ellipsis;
   white-space: nowrap;
   margin-bottom: @size-4;
-  border: 1px solid @color-contrast-dark-alternate;
+  border: 1px solid @color-secondary-action-outline;
   border-radius: @size-3;
 
   &:last-of-type {
@@ -34,7 +34,7 @@
 .Timestamp {
   .text-9;
 
-  color:@color-dark-alternate;
+  color:@color-inactive-text;
 }
 
 .Seen {
@@ -43,7 +43,7 @@
   > div {
     > div:first-of-type > div,
     > div:nth-of-type(2) > span {
-      color: @color-dark-alternate;
+      color: @color-inactive-text;
     }
   }
 }
@@ -61,7 +61,7 @@
 .Title {
   .text-10;
 
-  color: @color-contrast-dark;
+  color: @color-dark-grey;
   overflow: hidden;
   text-overflow: ellipsis;
   white-space: nowrap;
@@ -72,7 +72,7 @@
 .Status {
   .mono-9-bold;
 
-  color: @color-contrast-dark;
+  color: @color-dark-grey;
   text-transform: uppercase;
 }
 
@@ -109,7 +109,7 @@
     width: @size-8;
     height: size-8;
     > path {
-      fill: @color-dark-alternate;
+      fill: @color-inactive-text;
     }
   }
 }
@@ -117,7 +117,7 @@
 .Description {
   .text-9;
 
-  color: @color-dark-alternate;
+  color: @color-inactive-text;
   overflow: hidden;
   text-overflow: ellipsis;
   white-space: nowrap;
