+++ packages/augur-ui/src/modules/alerts/components/alerts-view.styles.less
@@ -34,7 +34,7 @@
 }
 
 .dark {
-  background-color: @color-contrast-dark;
+  background-color: @color-dark-grey;
 }
 
 .parent {
