+++ packages/augur-ui/src/modules/app/components/gas-price-edit.styles.less
@@ -94,5 +94,5 @@
 }
 
 .nonStandard {
-  color: @color-in-reporting;
+  color: @color-reporting;
 }
