+++ packages/augur-ui/src/modules/common/filter-notice.styles.less
@@ -34,7 +34,7 @@
   svg {
     height: @size-19;
     width: @size-19;
-    fill: @color-invalid-primary;
+    fill: @color-outcome-invalid;
   }
 
   rect, path {
