+++ packages/augur-ui/src/modules/common/form-styles.less
@@ -13,7 +13,7 @@
 .Form__error {
   .text-16;
 
-  color: @color-error-text;
+  color: @color-error;
   display: inline-block;
   margin-left: 2.32rem;
   margin-right: 1.5rem;
@@ -31,7 +31,7 @@
 .Form__error--even {
   .text-16;
 
-  color: @color-error-text;
+  color: @color-error;
   display: inline-block;
   margin-left: 0;
   margin-right: @size-24;
@@ -75,7 +75,7 @@
 }
 
 .Form__error--field {
-  border-color: @color-error-text;
+  border-color: @color-error;
   border-width: 1px;
 }
 
@@ -134,7 +134,7 @@
       &.error {
         .text-16;
 
-        color: @color-error-text;
+        color: @color-error;
       }
     }
 
@@ -146,7 +146,7 @@
       background-color: @color-alternate-info-text;
 
       &.error {
-        border-color: @color-error-text;
+        border-color: @color-error;
       }
 
       &::placeholder {
@@ -400,7 +400,7 @@ li.field--short {
   }
 
   .Form__error--field {
-    border-color: @color-error-text;
+    border-color: @color-error;
     border-width: 1px;
   }
