+++ packages/augur-ui/src/modules/common/labels.styles.less
@@ -227,7 +227,7 @@
     padding-left: @size-8;
 
     &.isAccented {
-      color: @color-error-text;
+      color: @color-error;
     }
   }
 
@@ -416,8 +416,8 @@ div.LinearPropertyViewTransaction {
 }
 
 .MarketStatus_reporting {
-  background: fade(@color-in-reporting, 15%);
-  color: @color-in-reporting;
+  background: fade(@color-reporting, 15%);
+  color: @color-reporting;
 }
 
 .MarketStatus_resolved {
@@ -432,7 +432,7 @@ div.LinearPropertyViewTransaction {
     margin: 0 @size-4;
 
     > path {
-      fill: @color-in-reporting;
+      fill: @color-reporting;
     }
   }
 }
@@ -494,7 +494,7 @@ div.LinearPropertyViewTransaction {
     > span {
       .mono-10-medium;
 
-      background: @color-error-text;
+      background: @color-error;
     }
   }
 }
@@ -723,7 +723,7 @@ div.LinearPropertyViewTransaction {
     }
 
     &.inReporting {
-      border-color: @color-in-reporting;
+      border-color: @color-reporting;
     }
 
     &.resolved {
