+++ packages/augur-ui/src/modules/common/progress.styles.less
@@ -35,7 +35,7 @@
 
   &.FirstBreakpoint {
     > span:last-of-type {
-      color: @color-in-reporting;
+      color: @color-reporting;
     }
   }
 
@@ -182,7 +182,7 @@
     height: @size-12;
 
     > span:first-of-type {
-      background: @color-in-reporting;
+      background: @color-reporting;
       height: @size-10;
       position: relative;
       top: 0;
@@ -237,7 +237,7 @@
   > p {
     .mono-10;
 
-    color: @color-in-reporting;
+    color: @color-reporting;
     text-align: center;
 
     &:first-of-type {
