+++ packages/augur-ui/src/modules/common/switch-labels-group.styles.less
@@ -105,7 +105,7 @@
   &.active {
     > span {
       &::after {
-        background: @color-in-reporting;
+        background: @color-reporting;
       }
     }
   }
