+++ packages/augur-ui/src/modules/market/components/common/outcome-trading-indicator/outcome-trading-indicator.styles.less
