+++ packages/augur-ui/src/modules/market/components/common/outcome-trading-indicator/outcome-trading-indicator.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { UP, DOWN, NONE } from "modules/common/constants";
-import Styles from "modules/market/components/common/outcome-trading-indicator/outcome-trading-indicator.style.less";
+import Styles from "modules/market/components/common/outcome-trading-indicator/outcome-trading-indicator.styles.less";
 
 export default function OutcomeTradingIndicator({
   tradingIndicator,
