+++ packages/augur-ui/src/modules/market/components/market-orders-positions-table/open-orders-table.styles.less
