+++ packages/augur-ui/src/modules/modal/components/common/common.styles.less
@@ -78,7 +78,7 @@
 
   form.ModalTightForm > p.Error,
   > p.Error {
-    color: @color-error-text;
+    color: @color-error;
   }
 
   form.ModalTightForm > p.Error {
@@ -224,7 +224,7 @@
 }
 
 .Error {
-  color: @color-error-text;
+  color: @color-error;
   display: flex;
   padding: 0;
 
@@ -239,7 +239,7 @@
 div.ErrorField,
 div.ErrorField:focus-within,
 div.ErrorField:focus {
-  border: @size-2 solid @color-error-text;
+  border: @size-2 solid @color-error;
   margin-bottom: @size-8;
 }
