+++ packages/augur-ui/src/modules/modal/modal.styles.less
@@ -63,7 +63,7 @@
 
   &.Bright {
     > h1 {
-      color: @color-in-reporting;
+      color: @color-reporting;
     }
   }
 
@@ -471,7 +471,7 @@ div.ButtonsRow:last-of-type {
       .text-10;
 
       align-content: center;
-      color: @color-error-text;
+      color: @color-error;
       display: flex;
       flex-flow: row nowrap;
       margin-top: @size-8;
@@ -483,12 +483,12 @@ div.ButtonsRow:last-of-type {
         width: @size-12;
 
         > circle {
-          stroke: @color-error-text;
+          stroke: @color-error;
         }
 
         > rect,
         > path {
-          fill: @color-error-text;
+          fill: @color-error;
         }
       }
     }
