+++ packages/augur-ui/src/modules/reporting/common.styles.less
@@ -85,7 +85,7 @@
   }
 
   &.invalid {
-    background-color: fade(@color-invalid, 50);
+    background-color: fade(@color-outcome-invalid, 50);
   }
 
   &.Round {
